需求[#15298]: AR-20200706-08-0101 任务管理接口开发
    [修改内容]: AR-20200706-08-0101 任务管理接口开发
    [描述说明]:
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/task/controller/TCruisePlanAttrController.java
@@ -108,7 +108,7 @@ public class TCruisePlanAttrController {
                             @RequestParam(value = "exceptionType", required = false) Integer exceptionType,
                             @RequestParam(value = "robotId", required = false) Long robotId,
                             @RequestParam(value = "position", required = false) String position,
-                            @RequestParam(value = "algorithmId", required = false) Integer algorithmId,
+                            @RequestParam(value = "algorithmId", required = false) Long algorithmId,
                             @RequestParam(value = "inferadAnalyze", required = false) String inferadAnalyze,
                             @RequestParam(value = "irTempBox", required = false) String irTempBox,
                             @RequestParam(value = "createTime", required = false) Date createTime,
